==> outdatedCode/5.05_ConfustionTables_Final_test3_randFor.Rmd <==
---
title: "Confusion Tables"
author: "Jill Deines"
date: "March 23, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Accuracy assessment via confusion tables for test 3, random forest, cleaned with one pass of a 3x3 majority filter, with interannual cleaning (pixels irrigated in a single year removed)

Confusion tables for the 2002 and 2015 test datasets. Jill's thinking is these are the final ones for the manuscript.

Tabulated for binary irrigated/nonirrigation classes. Only doing for both certainty levels (1 & 2) but the function can be altered to just use certainty level 1.


```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=FALSE, cache.path='cache/5.05_confusion/',
               fig.path='figure/5.05_confusion_final/')
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE}
library(rgdal)
library(caret)
library(knitr)

# source project helper functions
source('functions/prettyConfusion.R')
source('functions/testToConfusion_GEE.R')
```

**User Variables**
Filepaths for GEE exports (points sampling the classified maps). These files were created with the following scripts and data:

* final map accuracies (merged, cleaned)
  * GEE script 05.01_accuracyAssessment_interannual
  * Map asset: interannualCleaned/test3_randFor_cleaned1x_2016_allClasses
  * Points: full set of 2002 (1036) and 2015 (1248) points
  
* raw map accuracies
  * GEE script: 05.0_accuracyAssessment
  * Map assets: classifiedRRB + c(test3_randomForest_500, test3_randomForest_500_annual,
  test3_randomForest_500_monthly)
  * Points: full 2002 and 2015 point sets for the mosaic (500*) and annual
  * Points: truncated 2002 (911) and 2015 (830) for monthly classifier based on data
  

```{r userVars, message=FALSE, echo=TRUE}
# load data
dataDir <- 'C:/Users/deinesji/Google Drive/GEE_tableExports/Accuracy_PointsforConfusionTables_test3_randFor'

# point dataset filenames
p2002_final <- '2002_RRB_NE_confusionData_RRB_test3_randFor_cleanedInterannual1x_allClasses_final.csv'
p2002_annual <- '2002_RRB_NE_annual_confusionData_RRB_test3_randFor_uncleaned_500tree_allClasses_annual.csv'
p2002_monthly <- '2002_RRB_NE_month_confusionData_RRB_test3_randFor_uncleaned_500tree_allClasses_monthly.csv'
p2002_composite <- '2002_RRB_NE_annual_confusionData_RRB_test3_randFor_uncleaned_500tree_allClasses_composite.csv'


p2015_final <- '2015_RRB_confusionData_RRB_test3_randFor_cleanedInterannual1x_allClasses_final.csv'
p2015_annual <- '2015_RRB_annual_confusionData_RRB_test3_randFor_uncleaned_500tree_allClasses_annual.csv'
p2015_monthly <- '2015_RRB_month_confusionData_RRB_test3_randFor_uncleaned_500tree_allClasses_monthly.csv'
p2015_composite <- '2015_RRB_annual_confusionData_RRB_test3_randFor_uncleaned_500tree_allClasses_composite.csv'
```

## 2002
Test dataset restricted to NE portion of study area; drought year

```{r test2002}
# final classification maps, all points
final2002 <- testToConfusionGee(datadir = dataDir, dataname = p2002_final,
                                 bandName = 'b2002', certainty = 2)
kable(final2002, digits=1)

# composite raw maps, all points
raw2002 <- testToConfusionGee(datadir = dataDir, dataname = p2002_composite,
                                 bandName = 'classification', certainty = 2)
kable(raw2002, digits=1)



# annual classifier raw maps, all points
annual2002 <- testToConfusionGee(datadir = dataDir, dataname = p2002_annual,
                                 bandName = 'classification', certainty = 2)
kable(annual2002, digits=1)

# monthly classifier raw maps, monthly points
month2002 <- testToConfusionGee(datadir = dataDir, dataname = p2002_monthly,
                                 bandName = 'classification', certainty = 2)
kable(month2002, digits=1)
```


## 2015
Test dataset for full study area; wet year

```{r test2015}
# final classification maps, all points
final2015 <- testToConfusionGee(datadir = dataDir, dataname = p2015_final,
                                 bandName = 'b2015', certainty = 2)
kable(final2015, digits=1)

# composite raw maps, all points
raw2015 <- testToConfusionGee(datadir = dataDir, dataname = p2015_composite,
                                 bandName = 'classification', certainty = 2)
kable(raw2015, digits=1)



# annual classifier raw maps, all points
annual2015 <- testToConfusionGee(datadir = dataDir, dataname = p2015_annual,
                                 bandName = 'classification', certainty = 2)
kable(annual2015, digits=1)

# monthly classifier raw maps, monthly points
month2015 <- testToConfusionGee(datadir = dataDir, dataname = p2015_monthly,
                                 bandName = 'classification', certainty = 2)
kable(month2015, digits=1)
```

## Combine
get combined accuracy stats on final products

```{r combined}
# add points together
combined <- final2002[1:3,1:3] + final2015[1:3,1:3]

# get updated producer's accuracy
combined$ProducerAccuracy <- NA
combined$ProducerAccuracy[1] <- combined$irrigated[1]/combined$total[1] * 100
combined$ProducerAccuracy[2] <- combined$notirrigated[2]/combined$total[2] * 100

# get updated consumer's accuracy
irrCon <- combined$irrigated[1]/combined$irrigated[3] * 100
nonCon <- combined$notirrigated[2]/combined$notirrigated[3] * 100
all <- (combined$irrigated[1] + combined$notirrigated[2]) / combined$total[3] * 100
newrow <- c(irrCon, nonCon, NA, all)

combined <- rbind(combined, newrow)
row.names(combined)[4] <- 'ConsumerAccuracy'

kable(combined, digits=1)
```
